Modular components for tab and result container
