--- src/lib/Counter.svelte
@@ -1,10 +1,0 @@
-<script>
-  let count = $state(0)
-  const increment = () => {
-    count += 1
-  }
-</script>
-
-<button onclick={increment}>
-  count is {count}
-</button>
